requirements.txt
tensorflow-gpu==2.1.0
pandas==1.1.5
scikit-learn==0.24.2
bert4keras==0.10.7
h5py==2.10.0
scikit-learn==0.24.1